fix: moved logic out of heart icon
